+++ controllers/kafkatopic_controller_test.go
@@ -75,6 +75,7 @@ func TestKafkaTopicReconciler(t *testing.T) {
 		r.newAivenGeneratedClient = func(_, _, _ string) (avngen.Client, error) {
 			return avn, nil
 		}
+		r.jitter = nil // deterministic RequeueAfter
 
 		res, err := r.Reconcile(t.Context(), ctrlruntime.Request{
 			NamespacedName: types.NamespacedName{